% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aut.R
\name{aut}
\alias{aut}
\title{Alternate Uses Task (AUT)}
\usage{
aut(data, .by = NULL, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL} (default), all data will be treated as from one subject and there
will be no grouping columns in the value returned.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
An object with the same class as \code{data} contains following values:

\item{originality}{The originality score.}

\item{flexibility}{The flexibility score.}

\item{fluency}{The fluency score.}
}
\description{
The Alternate Uses Task (AUT) is a test of divergent thinking, which is a
component of creativity. The test-taker is given a common object and asked
to list as many uses for that object as they can think of. The test is
scored based on the number of uses generated, the originality of the uses,
the flexibility of the uses, and the fluency of the uses. The originality
score is based on the number of participants who generated the same use for
the object. The flexibility score is based on the number of categories of
uses generated. The fluency score is based on the total number of uses
generated.
}
