% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{bart}
\alias{bart}
\title{Calculates index scores for BART game specifically}
\usage{
bart(data, by, vars_input)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{by}{The column name(s) in \code{data} used to be grouped by.}

\item{vars_input}{The matched variable names for further processing. Typically setted automatically by \code{\link[=preproc_data]{preproc_data()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{mean_pumps}{Mean of hits for balloons not exploded.}
\item{mean_pumps_raw}{Mean of hits for all balloons.}
\item{num_explosion}{Number of exploded balloons.}
}
\description{
The adjusted and unadjusted BART scores are both returned.
}
