% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cong_eff.R
\name{calc_cong_eff}
\alias{calc_cong_eff}
\title{Congruency effect}
\usage{
calc_cong_eff(
  data,
  name_cong = "Type",
  name_rt = "rt_cor",
  name_acc = "acc_cor",
  values_cong = c("Congruent", "Incongruent")
)
}
\arguments{
\item{data}{A \code{data.frame} from which the indices are calculated.}

\item{name_cong}{The name of the variable in \code{data} storing the congruency
info.}

\item{name_rt}{The name of the variable in \code{data} storing reaction time data.}

\item{name_acc}{The name of the variable in \code{data} storing accuracy data.}

\item{values_cong}{The values of the congruency info variable, in which the
first is about 'congruent', and the second about 'incongruent'.}
}
\description{
Utility function to calculate indices related to congruency effect.
}
\keyword{internal}
