% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-speed-accuracy.R
\name{calc_ddm}
\alias{calc_ddm}
\title{Calculate DDM indices}
\usage{
calc_ddm(
  data,
  name_rt = "rt",
  name_acc = "acc",
  name_truth = NULL,
  rt_unit = c("ms", "s")
)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{name_rt}{The column name of the \code{data} input whose values are user's reaction time, in which is a \code{numeric} vector with \code{NA} values representing abnormal reaction times.}

\item{name_acc}{The column name of the \code{data} input whose values are user's correctness, in which is a \code{numeric} vector so coded that 1 means scoring correct, 0 means scoring incorrect, and that -1 means no response is made.}

\item{name_truth}{The condition which is depended upon to calculate different
drifting rates. Default is \code{NULL}. Must be a \code{\link[=factor]{factor()}}.}

\item{rt_unit}{The unit of response time in \code{data}.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains drifting rate for
different conditions, decision boundary and non-decision time.
}
\description{
There are a lot of papers proposed to use drift-diffusion model to model
speed accuracy data. Here we mainly focused on the drifting rate (v),
decision boundary (a) and non-decision time (t0).
}
\keyword{internal}
