% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_staircase_wetherill}
\alias{calc_staircase_wetherill}
\title{Calculate threshold by staircase method}
\usage{
calc_staircase_wetherill(x)
}
\arguments{
\item{x}{The levels in data.}
}
\value{
The mean threshold.
}
\description{
Here we used the method suggested by Wetherill et al (1966).
}
\keyword{internal}
