% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_switch_cost.R
\name{calc_switch_cost}
\alias{calc_switch_cost}
\title{Switch cost}
\usage{
calc_switch_cost(
  data,
  by,
  name_type_block,
  name_type_switch,
  name_rt,
  name_acc
)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{by}{The column name(s) in \code{data} used to be grouped by.}

\item{name_type_block}{The column name of the \code{data} input whose values are
the block type, in which is a \code{character} vector with "pure block" (value:
\code{"pure"}) and "mixed block" (value: \code{"mixed"}).}

\item{name_type_switch}{The column name of the \code{data} input whose values are
the switch type, in which is a \code{character} vector with at least \code{"switch"}
and \code{"repeat"} values. Other values if corresponding to \code{"pure"} block,
will be used as task names.}

\item{name_rt}{The column name of the \code{data} input whose values are user's reaction time, in which is a \code{numeric} vector with \code{NA} values representing abnormal reaction times.}

\item{name_acc}{The column name of the \code{data} input whose values are user's correctness, in which is a \code{numeric} vector so coded that 1 means scoring correct, 0 means scoring incorrect, and that -1 means no response is made.}
}
\description{
Utility function to calculate general and specific switch cost.
}
\keyword{internal}
