% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity.R
\name{capacity}
\alias{capacity}
\title{Visual Arrays}
\usage{
capacity(data, .by = NULL, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL} (default), all data will be treated as from one subject and there
will be no grouping columns in the value returned.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
An object with the same class as \code{data} contains following values:

\item{k}{The mean capacity in all conditions.}

\item{k3}{The capacity in condition of 3 arrays.}

\item{k5}{The capacity in condition of 5 arrays.}

\item{k7}{The capacity in condition of 7 arrays.}

\item{k9}{The capacity in condition of 9 arrays.}
}
\description{
The visual arrays task is used to measure working memory capacity. Here we
calculate the capacity from data. Note this is used when the whole visual
arrays are to be detected.
}
