% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_resp_metric.R
\name{check_resp_metric}
\alias{check_resp_metric}
\title{Check whether basic data response metrics meet requirements}
\usage{
check_resp_metric(
  x,
  grp,
  crit_resp_rate = 0.8,
  crit_acc = NULL,
  check_type = c("all", "resp_rate", "accuracy", "none"),
  size = NULL,
  chance = 0.5
)
}
\arguments{
\item{x}{A \code{numeric} vector so coded that 1 means scoring correct, 0 means
scoring incorrect, and that -1 means no response is made.}

\item{grp}{A \link[collapse:GRP]{GRP} object.}

\item{crit_resp_rate}{The required minimal valid response rate. Default to
0.8.}

\item{crit_acc}{The required minimal accuracy. Default to \code{NULL}, a minimal
accuracy was calculated based on \code{size} and \code{chance} argument (see
"Details").}

\item{check_type}{"all": check all metrics. "resp_rate": check response rate
only. "accuracy": check accuracy only. "none": skip all checking.}

\item{size}{The number of trials. Default to \code{NULL}, then get it from input
data.}

\item{chance}{The chance level of score correct for each trial. Default to
0.5, which is so for most games.}
}
\value{
A \code{data.frame} with columns of grouping and a new column named
\code{is_normal}, in which \code{TRUE} means that the response metrics of that group
meet the requirements, and \code{FALSE} means not. If the \code{check_type} is set as
"none", \code{NA} is returned.
}
\description{
This function is used to unify the check of response metrics for all the
data. Two basic metrics, i.e. valid response rate and total accuracy, are
checked (both can be opted out).
}
\details{
Minimal valid response rate is required for all the games that need users'
response for all the trials. Minimal accuracy is required for all the games
that the chance accuracy level can be acquired for each trial.

Minimal accuracy calculation is based on a statistical hypothesis testing
against the null hypothesis that the user responded at random. If so, the
number of trials a user scores correct will obey a Bernoulli distribution
\eqn{B(n,p)} with parameter \eqn{n} of the number of trials (set in \code{size}
argument) and \eqn{p} of the chance accuracy level (set in \code{chance}
argument). So the minimal accuracy can be calculated based on the 95\% region
of rejection.
}
\keyword{internal}
