% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexswitch.R
\name{complexswitch}
\alias{complexswitch}
\title{Complex switch: congruence and switch combined}
\usage{
complexswitch(data, by, vars_input)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{by}{The column name(s) in \code{data} used to be grouped by.}

\item{vars_input}{The matched variable names for further processing. Typically set by the high-level function \code{\link[=preproc_data]{preproc_data()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the following variables:
\item{mrt_con}{Mean reaction time for congruent trials.}
\item{mrt_inc}{Mean reaction time for incogruent trials.}
\item{cong_eff_rt}{Congruence effect of reaction time (RT), i.e.,
incongruent RT - congruent RT.}
\item{pc_con}{Percent of correct for congruent trials.}
\item{pc_inc}{Percent of correct for incogruent trials.}
\item{cong_eff_pc}{Congruency effect of percent of correct (PC), i.e.,
congruent PC - incongruent PC.}
\item{mrt_pure}{Mean reaction time for non-mixed blocks.}
\item{mrt_repeat}{Mean reaction time for repeat trials.}
\item{mrt_switch}{Mean reaction time for switch trials.}
\item{switch_cost_rt_gen}{General switch cost (based on mean reaction
times).}
\item{switch_cost_rt_spe}{Specific switch cost (based on mean reaction
times).}
\item{pc_pure}{Percent of correct for non-mixed blocks.}
\item{pc_repeat}{Percent of correct for repeat trials.}
\item{mrt_switch}{Percent of correct for switch trials.}
\item{switch_cost_pc_gen}{General switch cost (based on percent of
correct).}
\item{switch_cost_pc_spe}{Specific switch cost (based on percent of
correct).}
}
\description{
Congruence effect and switch cost both exist in these tasks. This function
just outputs scores of these two effects.
}
\seealso{
\code{\link[=congeff]{congeff()}} and \code{\link[=switchcost]{switchcost()}} for congruence effect and switch cost
respectively.
}
