% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_abnormal_output.R
\name{compose_abnormal_output}
\alias{compose_abnormal_output}
\title{Compose output containing \code{NA}s with names required}
\usage{
compose_abnormal_output(vars_output)
}
\arguments{
\item{vars_output}{Variable names in the output}
}
\value{
Use \code{NA_real} as values. And a variable of \code{FALSE} value is added.
}
\description{
Compose output containing \code{NA}s with names required
}
\keyword{internal}
