% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condstairs.R
\name{condstairs}
\alias{condstairs}
\title{Adaptive Filtering}
\usage{
condstairs(data, .by = NULL, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL} (default), all data will be treated as from one subject and there
will be no grouping columns in the value returned.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
An object with the same class as \code{data} contains following values:

\item{capacity0}{The mean number of target in condition of no distractors.}

\item{capacity2}{The mean number of target in condition of 2 distractors.}

\item{capacity}{The mean number of target in both conditions.}

\item{efficiency}{The filtering efficiency, .i.e, difference between
condition of no distractors and two distractors.}
}
\description{
This is an self-adaptive version (item number is adaptive to user's ability)
of filtering task. Only two conditions are included, i.e., condition of no
distractor and condition of two distractors.
}
