% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drm.R
\name{drm}
\alias{drm}
\title{Deese-Roediger-McDermott (DRM) paradigm}
\usage{
drm(data, .by = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to \code{NULL} (default), all data will be treated as from one subject and there will be no grouping columns in the value returned.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{tm_dprime}{Sensitivity (d') of true memory (against "foil" stimuli).}
\item{tm_bias}{Bias of true memory (against "foil" stimuli).}
\item{fm_dprime}{Sensitivity (d') of false memory.}
\item{fm_bias}{ias of false memory.}
}
\description{
This is a classical false memory test. Here calculates the effect size of
false memory.
}
