% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ddm.R
\name{fit_ddm}
\alias{fit_ddm}
\title{Fit a Drift Diffusion Model}
\usage{
fit_ddm(
  data,
  name_rt = "rt",
  name_acc = "acc",
  name_truth = NULL,
  rt_unit = c("ms", "s"),
  n_fit = 5,
  seed = 1
)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{name_rt}{The column name of the \code{data} input whose values are user's reaction time, in which is a \code{numeric} vector with \code{NA} values representing abnormal reaction times.}

\item{name_acc}{The column name of the \code{data} input whose values are user's correctness, in which is a \code{numeric} vector so coded that 1 means scoring correct, 0 means scoring incorrect, and that -1 means no response is made.}

\item{name_truth}{The condition which is depended upon to calculate different
drifting rates. Default is \code{NULL}. Must be a \code{\link[=factor]{factor()}}.}

\item{rt_unit}{The unit of response time in \code{data}.}

\item{n_fit}{Number of fits to try to find the best estimate.}

\item{seed}{Random seed. Default is 1 so that results can be reproduced.}
}
\value{
A \code{\link[=list]{list()}} with structure the same as \code{\link[=optim]{optim()}}.
}
\description{
Gain drift diffusion model for different conditions (if specified).
}
