% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_numerosity.R
\name{fit_numerosity}
\alias{fit_numerosity}
\title{Fit a Simple Numerosity Model}
\usage{
fit_numerosity(data, name_bigset, name_smallset, name_acc, n_fit = 5, seed = 1)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{name_bigset, name_smallset}{Variable name in \code{data} indicates bigger and
smaller set.}

\item{name_acc}{Variable name in \code{data} indicates user's response is correct
or not.}

\item{n_fit}{Number of fits to try to find the best estimate.}

\item{seed}{Random seed. Default is 1 so that results can be reproduced.}
}
\value{
A \code{\link[=list]{list()}} with structure the same as \code{\link[=optim]{optim()}}.
}
\description{
This model assumes the distribution of mental representation for a given
number/count k is N(k, (w * k) ^ 2).
}
