% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fname.R
\name{fname}
\alias{fname}
\title{Face Name Task}
\usage{
fname(data, .by = NULL, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL} (default), all data will be treated as from one subject and there
will be no grouping columns in the value returned.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
An object with the same class as \code{data} contains following values:

\item{fnn}{The overall number of correct face name pairs.}

\item{fno}{The overall number of correct face occupation pairs.}

\item{fntotal}{The overall number of correct face name and occupation
pairs.}
}
\description{
This is used to do face name task indicator calculations. Current version
integrates a occupation memory task.
}
\details{
Following Amariglio et al. (2012), we will calculate three scores: FN-N, FN-O
and FN-Total.
}
\references{
Amariglio, R. E., Frishe, K., Olson, L. E., Wadsworth, L. P.,
Lorius, N., Sperling, R. A., & Rentz, D. M. (2012). Validation of the Face
Name Associative Memory Exam in cognitively normal older individuals.
Journal of Clinical and Experimental Neuropsychology, 34(6), 580–587.
https://doi.org/10.1080/13803395.2012.666230
}
