% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igt.R
\name{igt}
\alias{igt}
\title{Iowa Gambling Task (modified)}
\usage{
igt(data, .by = NULL, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL} (default), all data will be treated as from one subject and there
will be no grouping columns in the value returned.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
An object with the same class as \code{data} contains following values:

\item{sum_outcome}{The total outcome over all trials.}

\item{perc_good}{The number of choices on "good" pools.}
}
\description{
A classical test on decision making. Read more details on
\href{https://en.wikipedia.org/wiki/Iowa_gambling_task}{wikipedia}. This modified
version uses pools to simulate cards, but the essential ideas are the same.
}
