% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlo.R
\name{jlo}
\alias{jlo}
\title{Judgment of Line Orientation}
\usage{
jlo(data, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{nc}{Count of correct responses.}
\item{mean_ang_err}{Mean of the response angle errors.}
\item{mean_log_err}{Mean of the log-transformed (of base 2) response angle
errors.}
}
\description{
This test is about visuo-spatial skills. For more details, read \href{https://en.wikipedia.org/wiki/Judgment_of_Line_Orientation}{this introduction}.
}
