% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nback.R
\name{nback}
\alias{nback}
\alias{dualnback}
\title{N Back Paradigm}
\usage{
nback(data, .by = NULL, .input = NULL, .extra = NULL)

dualnback(data, .by = NULL, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL} (default), all data will be treated as from one subject and there
will be no grouping columns in the value returned.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
An object with the same class as \code{data} contains following values (tripled
for dual n-back):

\item{pc}{Percentage of correct responses.}

\item{mrt}{Mean reaction time.}

\item{ies}{Inverse efficiency score.}

\item{rcs}{Rate correct score.}

\item{lisas}{Linear integrated speed-accuracy score.}

\item{dprime}{Sensitivity index.}
}
\description{
A classical working memory test.
}
\details{
Note for dual n-back, the returned indices include those for each modal and
both modals so they are tripled with additional suffix after each index name.
}
