% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nle.R
\name{nle}
\alias{nle}
\title{Number Line Estimation}
\usage{
nle(data, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{mean_abs_err}{Mean absolute error.}
\item{mean_log_err}{Mean log absolute error.}
}
\description{
A classical test on subject's numerical estimation skills.
}
