% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_char_acc.R
\name{parse_char_acc}
\alias{parse_char_acc}
\title{Convert character accuracy}
\usage{
parse_char_acc(x, delim = "-")
}
\arguments{
\item{x}{The character to be parsed.}

\item{delim}{Delimiter used to join correctness when forming the character.
Usually is hyphen (i.e., \code{"-"}), which is the default.}
}
\value{
A numeric vector of the parsed correctness. Might produce \code{NA}s.
}
\description{
Simple function converts character correctness to numeric one.
}
\keyword{internal}
