% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{parse_char_resp}
\alias{parse_char_resp}
\title{Convert character responses}
\usage{
parse_char_resp(x, delim = "-", convert_numeric = TRUE)
}
\arguments{
\item{x}{The character vector to be parsed.}

\item{delim}{Delimiter used to join correctness when forming the character.
Usually is hyphen (i.e., \code{"-"}), which is the default.}

\item{convert_numeric}{A logical value indicating if the values should be
converted to \code{numeric} ones.}
}
\value{
A list of the parsed result, the same length as the input vector.
}
\description{
Simple function converts character correctness to numeric one.
}
\keyword{internal}
