% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.iquizoo-package.R
\docType{package}
\name{preproc.iquizoo}
\alias{preproc.iquizoo-package}
\alias{preproc.iquizoo}
\title{Utility functions for data processing for IQUIZOO games.}
\description{
Several couples of games are developed by IQUIZOO.COM. Here are the functions used to do data processing for all of those games.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://iquizoo.github.io/preproc.iquizoo/}
  \item \url{https://github.com/iquizoo/preproc.iquizoo}
  \item Report bugs at \url{https://github.com/iquizoo/preproc.iquizoo/issues}
}

}
\author{
\strong{Maintainer}: Liang Zhang \email{psychelzh@outlook.com} (\href{https://orcid.org/0000-0001-9041-1150}{ORCID})

}
\keyword{internal}
