% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{preproc}
\alias{preproc}
\title{Call low-level functions given data}
\usage{
preproc(data, .fn, .by = NULL, ..., character.only = FALSE)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.fn}{The name of a function, given as a \link[base:name]{name} or literal
character string, depending on whether \code{character.only} is \code{FALSE}
(default) or \code{TRUE}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to \code{NULL} (default), all data will be treated as from one subject and there will be no grouping columns in the value returned.}

\item{...}{These dots are for future extensions and must be empty.}

\item{character.only}{A logical indicating whether \code{.fn} can be assumed to be
character strings.}
}
\value{
A \link[tibble:tibble-package]{tibble} of game performances returned by
low-level functions. Attributes will be the same with \code{data}.
}
\description{
This is the really "entry point" function that gives a convenient access to
all the low-level pre-processing functions. These low-level functions are
also exported and documented with the details of the returned values.
}
\author{
Liang Zhang \href{mailto:psychelzh@outlook.com}{psychelzh@outlook.com}
}
