% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racer.R
\name{racer}
\alias{racer}
\title{NeuroRacer Modified}
\usage{
racer(data, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:

\item{mean_pro_ratio}{Mean protection duration ratio.}

\item{dprime}{Sensitivity index of detection task.}
}
\description{
This is a modified version of NeuroRacer game.
}
