% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapm.R
\name{rapm}
\alias{rapm}
\title{Raven's Advanced Progressive Matrices}
\usage{
rapm(data, .by = NULL, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL} (default), all data will be treated as from one subject and there
will be no grouping columns in the value returned.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
An object with the same class as \code{data} contains following values:

\item{nc_prac}{Number of correct items for set I.}

\item{nc_test}{Number of correct items for set II.}

\item{nc_total}{Number of correct items for whole set.}
}
\description{
This test contains two sets, namely set I and set II, and set I is a practice
set, whereas set II is the test set. So scores for each set and whole set are
calculated here.
}
