% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refframe.R
\name{refframe}
\alias{refframe}
\title{Spatial Reference Frame}
\usage{
refframe(data, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{mean_dist_err_allo/mean_dist_err_ego}{Mean of the response distance
errors for allocentric and egocentric conditions respectively.}
\item{mean_log_err_allo/mean_log_err_ego}{Mean of the log-transformed (of
base \eqn{e}) response distance errors for allocentric and egocentric
conditions respectively.}
}
\description{
Typically, two classes of spatial frames of reference: "egocentric" and
"allocentric". The spatial acuity for both classes are calculated.
}
