% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinf.R
\name{reinf}
\alias{reinf}
\title{Probability Reinforcement Learning}
\usage{
reinf(data, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:

\item{pc_approach}{The percent of correct for approach trials.}

\item{pc_avoid}{The percent of correct for avoid trials.}

\item{pc_learn}{The total percent of correct in the learn phase.}

\item{pc_test}{The total percent of correct in the test phase.}
}
\description{
A classical reinforcement learning test.
}
