% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\name{rt}
\alias{rt}
\alias{crt}
\alias{srt}
\title{Reaction Times}
\usage{
crt(data, .input = NULL, .extra = NULL)

srt(data, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:

\item{mrt}{Mean reaction time.}

\item{rtsd}{Standard deviation of reaction times.}

\item{nc}{Count of correct responses. Only for \code{\link[=crt]{crt()}}.}

\item{ies}{Inverse efficiency score. Only for \code{\link[=crt]{crt()}}.}

\item{rcs}{Rate correct score. Only for \code{\link[=crt]{crt()}}.}

\item{lisas}{Linear integrated speed-accuracy score. Only for \code{\link[=crt]{crt()}}.}

\item{v}{Drifting rate of ddm model. Only for \code{\link[=crt]{crt()}}.}

\item{a}{Threshold separation of ddm model. Only for \code{\link[=crt]{crt()}}.}

\item{t0}{Non-decision time of ddm model. Only for \code{\link[=crt]{crt()}}.}
}
\description{
Choice Reaction Time (CRT) and Simple Reaction Time (SRT) are classical tests
of human reaction times. These functions calculates the mean and standard
deviation of reaction times. In addition, subjects can commit errors in CRT
tests, so the number of correct responses is also calculated in \code{\link[=crt]{crt()}}.
}
