% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staircase.R
\name{staircase}
\alias{staircase}
\title{Threshold estimation from staircase method}
\usage{
staircase(data, .by = NULL, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL} (default), all data will be treated as from one subject and there
will be no grouping columns in the value returned.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:

\item{thresh_peak_valley}{The mean threshold of peaks and valleys.}

\item{thresh_last_block}{The mean threshold of the last block.}
}
\description{
A very simple method is used here, i.e., averaging all the levels in the last
block.
}
\details{
This is under the assumption that the last block is stable enough so that an
average of the levels is just the convergence of the threshold.
}
