% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopsignal.R
\name{stopsignal}
\alias{stopsignal}
\title{Calculates index scores for stop signal game}
\usage{
stopsignal(data, ...)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{...}{Other input argument for future expansion.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the following variables:
\item{pc_all}{Percent of correct for all the responses.}
\item{pc_go}{Percent of correct for the go trials only.}
\item{medrt_go}{Median reaction time (ms) of go trials.}
\item{mean_ssd}{Mean stop signal delay time (ms).}
\item{ssrt}{Stop signal reaction time (ms).}
\item{is_normal}{Checking result whether the data is normal.}
}
\description{
The main index for stop signal games is the so-called stop signal reaction
time (\emph{SSRT}). Here the method used is the same as that in stair-case method.
}
