% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synwin.R
\name{synwin}
\alias{synwin}
\title{SynWin Test}
\usage{
synwin(data, .by = NULL, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL} (default), all data will be treated as from one subject and there
will be no grouping columns in the value returned.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
An object with the same class as \code{data} contains following values:

\item{score_total}{Total score. Sum of the three sub-tests.}

\item{score_mem}{Score in the memory sub-test.}

\item{score_vis}{Score in visual monitoring sub-test.}

\item{score_aud}{Score in auditory monitoring sub-test.}
}
\description{
This is a multi-task game designed by Elsmore (1994).
}
