% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{update_settings}
\alias{update_settings}
\title{Update settings with option settings}
\usage{
update_settings(origin, option)
}
\arguments{
\item{origin}{The original settings.}

\item{option}{The name in \code{\link[=options]{options()}} to read for updating.}
}
\value{
An update list of settings.
}
\description{
Options are set in list can be tricky to update. This function makes partly
adding custom options work.
}
\keyword{internal}
