% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg.R
\name{vg}
\alias{vg}
\title{Verb Generation}
\usage{
vg(data, .by = NULL, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL} (default), all data will be treated as from one subject and there
will be no grouping columns in the value returned.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
An object with the same class as \code{data} contains following values:

\item{score}{Average cosine distance between the generated verbs and the
given nouns.}
}
\description{
This test requires the user to generate a verb that is semantically related
to each given noun. The score is the average cosine distance between the
generated verb and the given nouns.
}
