% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{bart}
\alias{bart}
\title{Balloon Analogue Risk Task}
\usage{
bart(data, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{mean_pumps}{Mean of hits for balloons not exploded.}
\item{mean_pumps_raw}{Mean of hits for all balloons.}
\item{num_explosion}{Number of exploded balloons.}
}
\description{
This task is deemed as a measure of impulsivity. Read more details on
\href{http://www.impulsivity.org/measurement/BART}{this website}.
}
