% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{bart}
\alias{bart}
\title{Calculates index scores for BART game specifically}
\usage{
bart(data, by = "id", vars_input = NULL, keep.by = TRUE)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{by}{The column(s) variable names in \code{data} used to be grouped by. Note,
the check of existence is in the higher level function, so here in this
function does not check them.}

\item{vars_input}{This is done by other functions, storing the matched
variable names for further processing.}

\item{keep.by}{A logical value. Should the grouping variable be kept in the
output?}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{mean_pumps}{Mean of hits for balloons not exploded.}
\item{mean_pumps_raw}{Mean of hits for all balloons.}
\item{num_explosion}{Number of exploded balloons.}
\item{is_normal}{Checking result whether the data is normal.}
}
\description{
The adjusted and unadjusted BART scores are both returned.
}
