% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{bart}
\alias{bart}
\title{Balloon Analogue Risk Task}
\usage{
bart(data, .by = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to \code{NULL} (default), all data will be treated as from one subject and there will be no grouping columns in the value returned.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{mean_pumps}{Mean of hits for balloons not exploded.}
\item{mean_pumps_raw}{Mean of hits for all balloons.}
\item{num_explosion}{Number of exploded balloons.}
}
\description{
This task is deemed as a measure of impulsivity. Read more details on
\href{http://www.impulsivity.org/measurement/BART}{this website}.
}
