% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bps.R
\name{bps}
\alias{bps}
\title{Calculates index scores for Behavioral Pattern Separation (BPS) game}
\usage{
bps(data, vars_input, by)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{vars_input}{This is done by other functions, storing the matched
variable names for further processing.}

\item{by}{The column(s) variable names in \code{data} used to be grouped by. If
set to \code{NULL}, all data will be treated as from one subject.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{pc}{Percent of correct responses.}
\item{p_sim_foil}{Percent of similar responses for "foil" stimuli.}
\item{p_sim_lure}{Percent of similar responses for "lure" stimuli.}
\item{p_sim_target}{Percent of similar responses for "target" stimuli.}
\item{bps_score}{BPS score.}
\item{is_normal}{Checking result whether the data is normal.}
}
\description{
The index was developed by Stark et. al. (2013), named as "BPS score".
}
