% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calculation.R
\name{calc_sdt}
\alias{calc_sdt}
\title{Signal Detection Theory}
\usage{
calc_sdt(data, name_type = "Type", name_acc = "ACC", values_type = c("s", "n"))
}
\arguments{
\item{data}{Required. A \code{data.frame} from which the indices are calculated.}

\item{name_type}{The name of the variable in \code{data} storing the signal type
info.}

\item{name_acc}{The name of the variable in \code{data} storing accuracy data.}

\item{values_type}{The values of signal type. The first is 'signal', and the
second is 'noise'.}
}
\description{
Utility function to calculate indices related to signal detection theory,
e.g., d', c.
}
\keyword{internal}
