% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_spd_acc}
\alias{calc_spd_acc}
\title{Calculate basic speed and accuracy scores}
\usage{
calc_spd_acc(
  data,
  by = NULL,
  name_acc = "acc",
  name_rt = "rt",
  rt_rm_out = TRUE,
  rt_unit = c("ms", "s")
)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL}, all data will be treated as from one subject.}

\item{name_acc}{The column name of the \code{data} input whose values are user's correctness, in which is a \code{numeric} vector so coded that 1 means scoring correct, 0 means scoring incorrect, and that -1 means no response is made.}

\item{name_rt}{The column name of the \code{data} input whose values are user's reaction time, in which is a \code{numeric} vector with \code{NA} values representing abnormal reaction times.}

\item{rt_rm_out}{A logical value. Whether remove response time outlier or
not. Default to \code{TRUE.}}

\item{rt_unit}{The unit of response time in \code{data}.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains the required scores.
}
\description{
This function calculates reaction time and accuracy scores, which are very
basic to most tests.
}
\details{
A major part of behavior tests is the speed and accuracy metrics, which are
actually the only ones that any test collects. Based on these two metrics,
two basic groups of scores can be obtained. The first group contains the mean
and standard deviations of response times the number, and the second contains
the number and percent of correct responses.
}
\section{Note}{


Outlier removal is recommended and it is set as default in here. The
outlier detection method used here is the one recommended by Cousineau, D.,
& Chartier, S. (2010).
}

\keyword{internal}
