% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-switch-congruence.R
\name{calc_switch_cost}
\alias{calc_switch_cost}
\title{Switch cost}
\usage{
calc_switch_cost(data, by, name_switch, name_rt, name_acc)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL}, all data will be treated as from one subject.}

\item{name_switch}{The column name of the \code{data} input whose values are
the switch type, in which is a \code{character} vector with at least \code{"switch"}
and \code{"repeat"} values.}

\item{name_rt}{The column name of the \code{data} input whose values are user's reaction time, in which is a \code{numeric} vector with \code{NA} values representing abnormal reaction times.}

\item{name_acc}{The column name of the \code{data} input whose values are user's correctness, in which is a \code{numeric} vector so coded that 1 means scoring correct, 0 means scoring incorrect, and that -1 means no response is made.}
}
\description{
Utility function to calculate general and specific switch cost.
}
\keyword{internal}
