% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_outliers_rt}
\alias{check_outliers_rt}
\title{Outliers Detection for response time data}
\usage{
check_outliers_rt(x, threshold = 2.5)
}
\arguments{
\item{x}{A vector of input reaction time data.}

\item{threshold}{The threshold for determining whether a value is outlier or
not. Default is set at 2.5, which is best sample size dependent.}
}
\value{
A logical vector of the detected outliers.
}
\description{
This method is also called "transform" method, because it does a
transformation before applying z-score method.
}
\details{
This is based on Cousineau, D., & Chartier, S. (2010), which is said to be
suitable for reaction time data.
}
\keyword{internal}
