% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexswitch.R
\name{complexswitch}
\alias{complexswitch}
\title{Calculates index scores for complex switch games}
\usage{
complexswitch(data, vars_input, by)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{vars_input}{The matched variable names for further processing. Typically setted automatically by \code{\link[=preproc_data]{preproc_data()}}.}

\item{by}{The column name(s) in \code{data} used to be grouped by.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the following variables:
\item{mrt_con}{Mean reaction time for congruent trials.}
\item{mrt_inc}{Mean reaction time for incogruent trials.}
\item{cong_eff_rt}{Congruency effect of reaction time (RT), i.e., RT
incongruency - RT congruency.}
\item{pc_con}{Percent of correct for congruent trials.}
\item{pc_inc}{Percent of correct for incogruent trials.}
\item{cong_eff_pc}{Congruency effect of percent of correct (PC), i.e., PC
congruency - PC incongruency.}
\item{mrt_pure}{Mean reaction time for non-mixed blocks.}
\item{mrt_repeat}{Mean reaction time for repeat trials.}
\item{mrt_switch}{Mean reaction time for switch trials.}
\item{switch_cost_rt_gen}{General switch cost (based on mean reaction
times).}
\item{switch_cost_rt_spe}{Specific switch cost (based on mean reaction
times).}
\item{pc_pure}{Percent of correct for non-mixed blocks.}
\item{pc_repeat}{Percent of correct for repeat trials.}
\item{mrt_switch}{Percent of correct for switch trials.}
\item{switch_cost_pc_gen}{General switch cost (based on percent of
correct).}
\item{switch_cost_pc_spe}{Specific switch cost (based on percent of
correct).}
}
\description{
Complex switch game, just as the name implied, switch cost will be of
interest. For its complexity, congruency effect will be calculated, too. In
fact it is just a combination of \code{\link[=congeff]{congeff()}} and \code{\link[=switchcost]{switchcost()}}.
}
