% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congeff.R
\name{congeff}
\alias{congeff}
\title{Congruence effect}
\usage{
congeff(data, by, vars_input)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{by}{The column name(s) in \code{data} used to be grouped by.}

\item{vars_input}{The matched variable names for further processing. Typically set by the high-level function \code{\link[=preproc_data]{preproc_data()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{mrt_inc}{Mean reaction time for incogruent trials.}
\item{mrt_con}{Mean reaction time for congruent trials.}
\item{cong_eff_rt}{Congruence effect of reaction time (RT), i.e.,
incongruent RT - congruent RT.}
\item{pc_inc}{Percent of correct for incogruent trials.}
\item{pc_con}{Percent of correct for congruent trials.}
\item{cong_eff_pc}{Congruence effect of percent of correct (PC), i.e.,
congruent PC - incongruent PC.}
}
\description{
The stimuli for these tasks contain \emph{congruent} and \emph{incongruent} conditions.
Typically, subjects score correct more easily and respond more quickly on
\emph{congruent} ones, thus creating a congruence effect. This function calculates
scores of these effects.
}
