% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{correct_rt_acc}
\alias{correct_rt_acc}
\title{Correct reaction time and accuracy values}
\usage{
correct_rt_acc(
  data,
  ...,
  crit_rt = 100,
  correct_type = c("both", "rt", "acc", "none"),
  name_rt = "RT",
  name_acc = "ACC"
)
}
\arguments{
\item{data}{Required. Raw data, a \code{data.frame}.}

\item{crit_rt}{Minimal reaction time in milliseconds. Reaction times that are
equal or less than this will be treated as illegal.}

\item{correct_type}{Four types are supported. \code{"both"} (default) means that
both response time and accuracy are corrected, \code{"none"} means no
correction, and \code{"rt"}/\verb{"acc} means correct the corresponding one.}

\item{name_rt}{The variable name of reaction time in the input \code{data}.}

\item{name_acc}{The variable name of accuracy in the input \code{data}.}
}
\value{
The data is added with two new columns. \code{rt_cor}: the corrected
reaction time; \code{acc_cor}: the corrected accuracy. They can both disappear,
because if there is no such correction, the column is added only if the
corresponding variable exists in the original \code{data}.
}
\description{
Generally speaking, a response time that is no more than 100 milliseconds is
impossible for human beings, which is possibly accompanied by an error
operation. This function tried to correct that by setting reaction time to
\code{NA} and accuracy to \code{0} (can be opted out) for those trials.
}
\keyword{internal}
