% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countcorrect.R
\name{countcorrect}
\alias{countcorrect}
\title{Count correct responses}
\usage{
countcorrect(data, by, vars_input)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{by}{The column name(s) in \code{data} used to be grouped by.}

\item{vars_input}{The matched variable names for further processing. Typically set by the high-level function \code{\link[=preproc_data]{preproc_data()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{nc}{Count of correct responses.}
}
\description{
This function calculates the number of correct responses.
}
