% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countcorrect.R
\name{countcorrect}
\alias{countcorrect}
\alias{countcorrect2}
\title{Count Correct Responses}
\usage{
countcorrect(data, by, vars_input)

countcorrect2(data, by, vars_input)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{by}{The column name(s) in \code{data} used to be grouped by.}

\item{vars_input}{The matched variable names for further processing. Typically set by the high-level function \code{\link[=preproc_data]{preproc_data()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{nc/nc_cor}{Count of correct responses. \code{nc_cor} means a corrected
count (subtracting number of errors), which is returned by
\code{\link[=countcorrect2]{countcorrect2()}}.}
}
\description{
These functions calculates the number of correct responses. \code{\link[=countcorrect2]{countcorrect2()}}
subtracts the count of errors, which is not the case for \code{\link[=countcorrect]{countcorrect()}}.
}
