% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts}
\alias{counts}
\alias{countcorrect}
\alias{countcorrect2}
\alias{sumweighted}
\alias{sumscore}
\title{Count Correct Responses}
\usage{
countcorrect(data, .by = NULL, .input = NULL, .extra = NULL)

countcorrect2(data, .by = NULL, .input = NULL, .extra = NULL)

sumweighted(data, .by = NULL, .input = NULL, .extra = NULL)

sumscore(data, .by = NULL, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL} (default), all data will be treated as from one subject and there
will be no grouping columns in the value returned.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{nc}{Count of correct responses. For \code{\link[=countcorrect]{countcorrect()}}.}
\item{nc_cor}{Corrected count of correct responses (subtracting number of
errors). For \code{\link[=countcorrect2]{countcorrect2()}}.}
\item{nc_weighted}{Count of weighted correct responses. For
\code{\link[=sumweighted]{sumweighted()}}.}
\item{nc_score}{Sum of scores. For \code{\link[=sumscore]{sumscore()}}.}
}
\description{
These functions count the number of correct responses. \code{\link[=countcorrect]{countcorrect()}}
counts the correct responses regardless of errors, \code{\link[=countcorrect2]{countcorrect2()}}
subtracts the number of errors from number of correct responses,
\code{\link[=sumweighted]{sumweighted()}} counts the correct responses by giving a weight for different
responses, \code{\link[=sumscore]{sumscore()}} adds up the score for each response.
}
