% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt.R
\name{cpt}
\alias{cpt}
\title{Continuous Performance Test}
\usage{
cpt(data, .by = NULL, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL} (default), all data will be treated as from one subject and there
will be no grouping columns in the value returned.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{nc}{Count of correct responses.}
\item{mrt}{Mean reaction time of hits.}
\item{rtsd}{Standard deviation of reaction times of hits.}
\item{dprime}{Sensitivity (d').}
\item{c}{Bias index.}
\item{commissions}{Number of errors caused by action.}
\item{omissions}{Number of errors caused by inaction.}
}
\description{
Continuous Performance Test (CPT) is a classical test for attention. There
are many methods used to calculate the performance index of this task, and
here only includes those common ones.
}
