% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driving.R
\name{driving}
\alias{driving}
\title{Driving Test}
\usage{
driving(data, .by)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{still_ratio}{The ratio of still duration in yellow light state.}
}
\description{
A test measuring impulsivity originally developed .by Gardner et. al. (2005).
}
