% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locmem.R
\name{locmem}
\alias{locmem}
\alias{locmem2}
\title{Location Memory}
\usage{
locmem(data, .by, .input)

locmem2(data, .by, .input)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by.}

\item{.input}{The matched variable names for further processing. Typically set by the high-level function \code{\link[=preproc]{preproc()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{nc_loc}{Count of correct responses for location.}
\item{mean_dist_err}{Mean of the response distance errors.}
\item{mean_log_err}{Mean of the log-transformed (of base \eqn{e}) response
distance errors.}
\item{nc_order}{Count of correct responses for order. For \code{\link[=locmem2]{locmem2()}}
only.}
}
\description{
Several tests are based on subject's spatial acuity, so typically a distance
error is collected and scores are calculated based on that error. \code{\link[=locmem]{locmem()}}
deal with the distance condition only. \code{\link[=locmem2]{locmem2()}} deals with a special case
when the response order and distance both matter.
}
