% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/london.R
\name{london}
\alias{london}
\title{London Tower}
\usage{
london(data, .by = NULL, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL} (default), all data will be treated as from one subject and there
will be no grouping columns in the value returned.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
An object with the same class as \code{data} contains following values:

\item{prop_perfect}{Proportion of responses with minimal moves.}

\item{mrt_init}{Mean initial response time.}
}
\description{
A classical test on problem solving.
}
