% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/london.R
\name{london}
\alias{london}
\title{London Tower}
\usage{
london(data, by, vars_input)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{by}{The column name(s) in \code{data} used to be grouped by.}

\item{vars_input}{The matched variable names for further processing. Typically set by the high-level function \code{\link[=preproc_data]{preproc_data()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{total_score}{Total score defined by the game itself.}
\item{mean_level}{Mean level reached.}
\item{level_score}{Sum of mean score (a ratio) for each level.}
}
\description{
A classical test on problem solving.
}
