% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_data_vars.R
\name{match_data_vars}
\alias{match_data_vars}
\title{Get the matched variables for input data}
\usage{
match_data_vars(data, vars_config)
}
\arguments{
\item{data}{Required. Raw data, a \code{data.frame}.}

\item{vars_config}{Required. A \code{data.frame} containing the configuration of
required variables. Possible variables are:
\itemize{
\item \code{field}: Required. Name references of all data variable names.
\item \code{name}: Required. Data variable names, in which lists all the possible
names for current data variable.
\item \code{type}: Optional. Can be "required" (must exist) or "optional" (try to
match, if not find, \code{NA} is returned and it must be handled in another
place). If this variable is not specified, it will default to "required".
}}
}
\description{
Input data require several variables to be existed, and some variables can
have alternatives. Here we check if all the variables exist, and if only one
of the alternatives exists for each variable. If the check is passed, the
matched result (a character vector) will be returned, or else a \code{NULL} value
will be returned.
}
\keyword{internal}
