% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_data_vars.R
\name{match_data_vars}
\alias{match_data_vars}
\title{Get the matched variables for input data}
\usage{
match_data_vars(data, prep_fun_name)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{prep_fun_name}{The name of the called pre-processing function.}
}
\value{
A \code{list} of the matched variables, using \code{NA} value to represent not
found, and using the names configured in the internal data.
}
\description{
Input data require several variables to be existed, and some variables can
have alternatives. Here we check if all the variables exist, and if only one
of the alternatives exists for each variable. If the check is passed, the
matched result (a character vector, \code{NA} means not found) will be returned.
}
\keyword{internal}
