% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nback.R
\name{nback}
\alias{nback}
\title{N Back Paradigm}
\usage{
nback(data, .by)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{pc}{Percentage of correct responses.}
\item{mrt}{Mean reaction time.}
\item{dprime}{Sensitivity index.}
\item{c}{Bias.}
}
\description{
A classical working memory test.
}
