% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nback.R
\name{nback}
\alias{nback}
\alias{dualnback}
\title{N Back Paradigm}
\usage{
nback(data, .input = NULL, .extra = NULL)

dualnback(data, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values (tripled
for dual n-back):
\item{pc}{Percentage of correct responses.}
\item{mrt}{Mean reaction time.}
\item{dprime}{Sensitivity index.}
}
\description{
A classical working memory test.
}
\details{
Note for dual n-back, the returned indices include those for each modal and
both modals so they are tripled with additional suffix after each index name.
}
