% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nback.R
\name{nback}
\alias{nback}
\title{N Back Paradigm}
\usage{
nback(data, by, vars_input)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{by}{The column name(s) in \code{data} used to be grouped by.}

\item{vars_input}{The matched variable names for further processing. Typically set by the high-level function \code{\link[=preproc_data]{preproc_data()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{pc}{Percentage of correct responses.}
\item{mrt}{Mean reaction time.}
\item{dprime}{Sensitivity index.}
\item{c}{Bias.}
}
\description{
A classical working memory test.
}
