% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nle.R
\name{nle}
\alias{nle}
\title{Number Line Estimation}
\usage{
nle(data, .by, .input)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by.}

\item{.input}{The matched variable names for further processing. Typically set by the high-level function \code{\link[=preproc]{preproc()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{mean_abs_err}{Mean absolute error.}
\item{mean_log_err}{Mean log absolute error.}
}
\description{
A classical test on subject's numerical estimation skills.
}
