% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsymncmp.R
\name{nsymncmp}
\alias{nsymncmp}
\title{Non-symbolic Number Comparison}
\usage{
nsymncmp(data, .by, .input)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by.}

\item{.input}{The matched variable names for further processing. Typically set by the high-level function \code{\link[=preproc]{preproc()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{pc}{Percentage of correct responses.}
\item{mrt}{Mean reaction time.}
\item{w}{Weber fraction.}
}
\description{
A classical test on subject's counting estimation skills.
}
\seealso{
\code{\link[=symncmp]{symncmp()}} for symbolic number comparison.
}
