% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{preproc.iquizoo}
\alias{preproc.iquizoo}
\alias{preproc.iquizoo-package}
\title{Utility functions for data processing for IQUIZOO games.}
\description{
Several couples of games are developed by IQUIZOO.COM. Here are the functions used to do data processing for all of those games.
}
\section{Package options}{


Preproc.iquizoo uses the following \code{\link[=options]{options()}} to configure behaviour:

\itemize{
\item \code{preproc.input}: input variable names, a named list of characters.

\item \code{preproc.output}: output variables names.

\item \code{preproc.extra}: extra options such as values for certain input, a
named list of characters.

}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://psychelzh.github.io/preproc.iquizoo}
  \item \url{https://github.com/psychelzh/preproc.iquizoo}
  \item Report bugs at \url{https://github.com/psychelzh/preproc.iquizoo/issues}
}

}
\author{
\strong{Maintainer}: Liang Zhang \email{psychelzh@outlook.com} (\href{https://orcid.org/0000-0001-9041-1150}{ORCID})

}
\keyword{internal}
