% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_data.R
\name{preproc_data}
\alias{preproc_data}
\title{Call low-level functions given data}
\usage{
preproc_data(data, prep_fun_name, by = NULL, ..., character.only = FALSE)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{prep_fun_name}{The name of a function, given as a \link[base:name]{name} or
literal character string, depending on whether \code{character.only} is \code{FALSE}
(default) or \code{TRUE}. Only functions from this package is supported for now.}

\item{by}{The column name(s) in \code{data} used to be grouped by. If set to \code{NULL} (default), all data will be treated as from one subject and there will be no grouping columns in the value returned.}

\item{...}{These dots are for future extensions and must be empty.}

\item{character.only}{A logical indicating whether \code{prep_fun_name} can be
assumed to be character strings.}
}
\value{
A \link[tibble:tibble-package]{tibble} of game performances returned by
low-level functions.
}
\description{
This is the really "entry point" function that gives a convenient access to
all the low-level pre-processing functions. These low-level functions are
also exported and documented with the details of the returned values.
}
\author{
Liang Zhang \href{mailto:psychelzh@outlook.com}{psychelzh@outlook.com}
}
