% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_data.R
\name{preproc_data}
\alias{preproc_data}
\title{Call pre-processing function to get the performances of the given data}
\usage{
preproc_data(data, prep_fun, by = "id", ...)
}
\arguments{
\item{data}{Raw data to be pre-processed. Should be of \code{data.frame} class.}

\item{prep_fun}{The function to be called. Currently, it should be a
\code{symbol}, representing the function.}
}
\description{
This is the really "entry point" function that gives a convenient access to
all the pre-processing functions. The documentation of each of them can be
accessed to know its details of the returned values.
}
