% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racer.R
\name{racer}
\alias{racer}
\title{NeuroRacer Modified}
\usage{
racer(data, .by = NULL, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL} (default), all data will be treated as from one subject and there
will be no grouping columns in the value returned.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
An object with the same class as \code{data} contains following values:

\item{mean_score}{Mean overlap score.}

\item{dprime}{Sensitivity index of detection task.}
}
\description{
This is a modified version of NeuroRacer game.
}
