% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refframe.R
\name{refframe}
\alias{refframe}
\title{Spatial Reference Frame}
\usage{
refframe(data, by, vars_input)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{by}{The column name(s) in \code{data} used to be grouped by.}

\item{vars_input}{The matched variable names for further processing. Typically set by the high-level function \code{\link[=preproc_data]{preproc_data()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{mean_dist_err_allo/mean_dist_err_ego}{Mean of the response distance
errors for allocentric and egocentric conditions respectively.}
\item{mean_log_err_allo/mean_log_err_ego}{Mean of the log-transformed (of
base \eqn{e}) response distance errors for allocentric and egocentric
conditions respectively.}
}
\description{
Typically, two classes of spatial frames of reference: "egocentric" and
"allocentric". The spatial acuity for both classes are calculated.
}
