% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\name{rt}
\alias{rt}
\alias{crt}
\alias{srt}
\title{Reaction Times}
\usage{
crt(data, by, vars_input)

srt(data, by, vars_input)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{by}{The column name(s) in \code{data} used to be grouped by.}

\item{vars_input}{The matched variable names for further processing. Typically set by the high-level function \code{\link[=preproc]{preproc()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{mrt}{Mean reaction time.}
\item{rtsd}{Standard deviation of reaction times.}
\item{nc}{Count of correct responses. Only for \code{\link[=crt]{crt()}}.}
}
\description{
Choice Reaction Time (CRT) and Simple Reaction Time (SRT) are classical tests
of human reaction times. These functions calculates the mean and standard
deviation of reaction times. In addition, subjects can commit errors in CRT
tests, so the number of correct responses is also calculated in \code{\link[=crt]{crt()}}.
}
