% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\name{rt}
\alias{rt}
\alias{crt}
\alias{srt}
\title{Reaction Times}
\usage{
crt(data, .by = NULL, .input = NULL, .extra = NULL)

srt(data, .by = NULL, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL} (default), all data will be treated as from one subject and there
will be no grouping columns in the value returned.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
An object with the same class as \code{data} contains following values:

\item{nc}{Count of correct responses. Only for \code{\link[=crt]{crt()}}.}

\item{mrt}{Mean reaction time.}

\item{rtsd}{Standard deviation of reaction times.}

\item{ies}{Inverse efficiency score. Only for \code{\link[=crt]{crt()}}.}

\item{rcs}{Rate correct score. Only for \code{\link[=crt]{crt()}}.}

\item{lisas}{Linear integrated speed-accuracy score. Only for \code{\link[=crt]{crt()}}.}
}
\description{
Choice Reaction Time (CRT) and Simple Reaction Time (SRT) are classical tests
of human reaction times. These functions calculates the mean and standard
deviation of reaction times. In addition, subjects can commit errors in CRT
tests, so the number of correct responses is also calculated in \code{\link[=crt]{crt()}}.
}
