% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/span.R
\name{span}
\alias{span}
\title{Span (spatial or verbal)}
\usage{
span(data, .by)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:]
\item{nc}{Count of correct responses.}
\item{max_span}{Maximal span.}
\item{mean_span}{Mean span.}
}
\description{
There is a bunch of tests measuring working memory span or attention span.
}
