% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/span.R
\name{span}
\alias{span}
\title{Span (spatial or verbal)}
\usage{
span(data, .by = NULL, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL} (default), all data will be treated as from one subject and there
will be no grouping columns in the value returned.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:]
\item{nc}{Count of correct responses.}
\item{max_span}{Maximal span.}
\item{mean_span}{Mean span.}
}
\description{
There is a bunch of tests measuring working memory span or attention span.
}
