% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopsignal.R
\name{stopsignal}
\alias{stopsignal}
\title{Stop Signal Paradigm}
\usage{
stopsignal(data, .by = NULL, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by. If set to
\code{NULL} (default), all data will be treated as from one subject and there
will be no grouping columns in the value returned.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the following variables:
\item{pc_all}{Percent of correct for all the responses.}
\item{pc_go}{Percent of correct for the go trials only.}
\item{pc_stop}{Percent of correct for the stop trials only.}
\item{rt_nth}{Percentile go reaction time (ms) based on \code{pc_stop}.}
\item{ssrt}{Stop signal reaction time (ms).}
}
\description{
A classical test on inhibition skills. The index calculation is now based on
https://doi.org/10.7554/eLife.46323.
}
