% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopsignal.R
\name{stopsignal}
\alias{stopsignal}
\title{Stop Signal Paradigm}
\usage{
stopsignal(data, .by)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.by}{The column name(s) in \code{data} used to be grouped by.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the following variables:
\item{pc_all}{Percent of correct for all the responses.}
\item{pc_go}{Percent of correct for the go trials only.}
\item{medrt_go}{Median reaction time (ms) of go trials.}
\item{ssrt}{Stop signal reaction time (ms).}
}
\description{
A classical test on inhibition skills.
}
