% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch-congruence.R
\name{switch-congruence}
\alias{switch-congruence}
\alias{complexswitch}
\alias{congeff}
\alias{switchcost}
\title{Task Switching and Stroop-like paradigm}
\usage{
complexswitch(data, .input = NULL, .extra = NULL)

congeff(data, .input = NULL, .extra = NULL)

switchcost(data, .input = NULL, .extra = NULL)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{.input, .extra}{Each is a \code{\link[=list]{list()}} containing all the input variable
names and special values for certain variables. See more in the details
section.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the following variables:

For the total task:

\item{pc}{Percent of correct}.

\item{mrt}{Mean reaction time.}

For congruence effect and switch cost, the following indices will be
included (including diffs and value for each condition):

\item{pc}{Percent of correct}.

\item{mrt}{Mean reaction time.}

\item{ies}{Inverse efficiency score.}

\item{rcs}{Rate correct score.}

\item{lisas}{Linear integrated speed-accuracy score.}

\item{v}{Drifting rate of ddm model.}
}
\description{
In task switching paradigms, two types of tasks switch between each other, so
the "\emph{switch cost}" can be calculated (using \code{\link[=switchcost]{switchcost()}}). Similarly, in
Stroop-like tasks, stimuli are classified into two conditions (i.e.,
"congruent" and "incongruent"), so the "\emph{congruence effect}" can be
calculated (using \code{\link[=congeff]{congeff()}}). There are also special types of tests where
congruence effect and switch cost both exist, from which \code{\link[=complexswitch]{complexswitch()}}
calculates both.
}
