% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexswitch.R
\name{switch-congruence}
\alias{switch-congruence}
\alias{complexswitch}
\alias{congeff}
\alias{switchcost}
\title{Task Switch and Stroop-like paradigm}
\usage{
complexswitch(data, by, vars_input)

congeff(data, by, vars_input)

switchcost(data, by, vars_input)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{by}{The column name(s) in \code{data} used to be grouped by.}

\item{vars_input}{The matched variable names for further processing. Typically set by the high-level function \code{\link[=preproc_data]{preproc_data()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the following variables:

For congruence effect:
\item{mrt_con}{Mean reaction time for congruent trials.}
\item{mrt_inc}{Mean reaction time for incogruent trials.}
\item{cong_eff_rt}{Congruence effect of reaction time (RT), i.e.,
incongruent RT - congruent RT.}
\item{pc_con}{Percent of correct for congruent trials.}
\item{pc_inc}{Percent of correct for incogruent trials.}
\item{cong_eff_pc}{Congruency effect of percent of correct (PC), i.e.,
congruent PC - incongruent PC.}

For switch cost:
\item{mrt_pure}{Mean reaction time for non-mixed blocks.}
\item{mrt_repeat}{Mean reaction time for repeat trials.}
\item{mrt_switch}{Mean reaction time for switch trials.}
\item{switch_cost_rt_gen}{General switch cost (based on mean reaction
times).}
\item{switch_cost_rt_spe}{Specific switch cost (based on mean reaction
times).}
\item{pc_pure}{Percent of correct for non-mixed blocks.}
\item{pc_repeat}{Percent of correct for repeat trials.}
\item{pc_switch}{Percent of correct for switch trials.}
\item{switch_cost_pc_gen}{General switch cost (based on percent of
correct).}
\item{switch_cost_pc_spe}{Specific switch cost (based on percent of
correct).}
}
\description{
There are two types of tasks switching between each other for task swicth
paradigms, thus a switch cost can be calculated (using \code{\link[=switchcost]{switchcost()}}).
Similarly, the stimuli for Stroop-like tasks contain \emph{congruent} and
\emph{incongruent} conditions, thus a congruence effect can be calculated (using
\code{\link[=congeff]{congeff()}}). There are also special types of tests in which congruence
effect and switch cost both exist, and \code{\link[=complexswitch]{complexswitch()}} calculates both.
}
