% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symncmp.R
\name{symncmp}
\alias{symncmp}
\title{Symbolic Number Comparison}
\usage{
symncmp(data, by, vars_input)
}
\arguments{
\item{data}{Raw data of class \code{data.frame}.}

\item{by}{The column name(s) in \code{data} used to be grouped by.}

\item{vars_input}{The matched variable names for further processing. Typically set by the high-level function \code{\link[=preproc_data]{preproc_data()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} contains following values:
\item{pc}{Percentage of correct responses.}
\item{mrt}{Mean reaction time.}
\item{dist_eff}{Distance effect.}
}
\description{
Several values including percentage of correct responses (pc), mean reaction
time (mrt), distance effect (dist_effect) and adjusted distance effect
(dist_effect_cor).
}
\seealso{
\code{\link[=nsymncmp]{nsymncmp()}} for non-symbolic number comparison.
}
