% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_settings}
\alias{update_settings}
\title{Update settings with option settings}
\usage{
update_settings(origin, updates)
}
\arguments{
\item{origin}{The original settings.}

\item{updates}{The updates to settings}
}
\value{
An update list of settings.
}
\description{
Options are set in list can be tricky to update. This function makes partly
adding custom options work.
}
\keyword{internal}
